Moved cluster_service_integration_tests.rs into cluster module
