+++ crates/flashq-grpc/tests/cluster_integration_tests.rs
@@ -1,0 +1,5 @@
+pub mod test_utilities;
+
+mod cluster {
+    pub mod cluster_service_integration_tests;
+}
